fix(extract/cisco/json): match unparseable allow-list on the raw product name
The allow-list is built from a full-dataset survey of the actual productNames,
so exact (raw) matching is intended: a whitespace-variant of a known-malformed
entry is a genuinely new pattern that the hard-error is meant to surface for
re-triage, not silently skip. Revert the trimmed lookup. Also consolidate
TestConvertProductName into json_test.go.
